+++ src/lib/aggregations/types.ts
@@ -14,6 +14,7 @@ import { MaxAggregation } from './get-max'
 import { CardinalityAggregation } from './get-cardinality'
 import { MinAggregation } from './get-min'
 import { StatsBucketAggregation } from './get-stats-bucket'
+import { NestedAggregation } from './get-nested'
 
 export type AggregationList<TDocument extends Document> =
     | AvgAggregation<TDocument>
@@ -32,6 +33,7 @@ export type AggregationList<TDocument extends Document> =
     | CardinalityAggregation<TDocument>
     | CompositeAggregation<TDocument>
     | StatsBucketAggregation
+    | NestedAggregation<TDocument>
 
 export type AggregationsContainer<TDocument extends Document> = Record<string, Aggregations<TDocument>>