+++ src/lib/responses/index.ts
@@ -1,3 +1,4 @@
 export * from './search'
 export * from './cluster-health'
 export * from './missing-value-aggregation'
+export * from './nested-aggregation'