+++ src/test/module/index.ts
@@ -1,3 +1,4 @@
 export * from './home.document'
 export * from './enums'
 export * from './test.service'
+export * from './animal.document'