+++ src/test/scripts/es-seed.ts
@@ -23,9 +23,23 @@ readFile(path)
             await client.indices.delete({ index })
         }
 
+        await client.indices.create({ index })
+
+        await client.indices.putMapping({
+            index,
+            body: {
+                properties: {
+                    animals: {
+                        type: 'nested'
+                    }
+                }
+            }
+        })
+
         await client.bulk({ body: records })
 
-        console.log('Seeded `homes` with:', records.length, 'results.')
+        // note: on line 17 index is added to each record, which means there are twice as many results - therefore the length is divided by 2
+        console.log('Seeded `homes` with:', records.length / 2, 'results.')
     })
     .catch(error => {
         throw new Error(`Failed to load homes seed: ${error.message}`)